+++ packages/core/src/tests/utility/mockOpenAI.ts
@@ -1,7 +1,10 @@
 import { CallbackManager } from "../../callbacks/CallbackManager";
 import { OpenAIEmbedding } from "../../embeddings";
 import { globalsHelper } from "../../GlobalsHelper";
-import { LLMChatParamsBase, OpenAI } from "../../llm/LLM";
+import { OpenAI } from "../../llm/LLM";
+import { LLMChatParamsBase } from "../../llm/types";
+
+export const DEFAULT_LLM_TEXT_OUTPUT = "MOCK_TOKEN_1-MOCK_TOKEN_2";
 
 export function mockLlmGeneration({
   languageModel,
@@ -14,7 +17,7 @@ export function mockLlmGeneration({
     .spyOn(languageModel, "chat")
     .mockImplementation(
       async ({ messages, parentEvent }: LLMChatParamsBase) => {
-        const text = "MOCK_TOKEN_1-MOCK_TOKEN_2";
+        const text = DEFAULT_LLM_TEXT_OUTPUT;
         const event = globalsHelper.createEvent({
           parentEvent,
           type: "llmPredict",